Project 1b/requirements.txt
yfinance
pandas
numpy
matplotlib
pyyaml
